requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.4.2
pydantic-settings==2.0.3
python-dateutil==2.8.2
